Hides toolbar when scrolling on entry details
Reorganizes the layout file for entry details so that the toolbar hides when scrolling down and reappears when scrolling up.

The only way to have this new layout work was to have the SwipeActionView at the top level, so now the toolbar also moves with the content when swiping between entries.