Fix arrows to not show before content load and arrows to be black on light theme
The arrows would briefly show up on screen before the entry loaded, and now the scroll view has a background set so that the arrows cannot be seen before the content is loaded.

Second fix is that the arrows were always white, even on the LIGHT theme. This fixes that by making the arrows black when the light theme is currently being used.