Fix issue with transparent navigation in landscape orientation
The transparent nav when in landscape orientation was causing all sorts of elements to render underneath the nav. The drawer now properly offsets as well as the SwipeRefreshLayout refreshing indicator bar. Also cleaned up some code and got the drawer header to render at the correct height with/without transparent status bar.