Enable strict mode and fix perf issues
Strict mode allows you to see when stuff is running on UI thread that is long running (and various other issues that are related). This enables that and then fixes the major perf issues that were obvious once strict mode was turned on. All of the issues fixed are centered around trying to run disk reads off UI thread. This should dramatically reduce frame drops when scrolling on enties fragment.